+++ assertj/src/main/java/be/iffy/fv/assertj/ValidationAssert.java
@@ -2,6 +2,7 @@ package be.iffy.fv.assertj;
 
 import be.iffy.fv.Validation;
 import org.assertj.core.api.AbstractAssert;
+import org.assertj.core.api.Assertions;
 import org.assertj.core.api.ObjectAssert;
 
 import java.util.function.Supplier;
@@ -39,7 +40,7 @@ public class ValidationAssert<SELF extends ValidationAssert<SELF, VALIDATION, T>
      * Creates a new instance of {@link ValidationAssert}.
      *
      * @param actual the actual value.
-     
+
      * @return the created assertion object.
      */
     public static <T> ValidationAssert<?, Validation<T>, T> assertThatValidation(Validation<T> actual) {
@@ -50,7 +51,7 @@ public class ValidationAssert<SELF extends ValidationAssert<SELF, VALIDATION, T>
      * Asserts that the validation is valid.
      *
      * @param actual the actual value.
-     
+
      * @return a {@link ObjectAssert} with the value inside the Validation for further assertions.
      */
     public static <T> ObjectAssert<T> assertValid(Validation<T> actual) {
@@ -62,7 +63,7 @@ public class ValidationAssert<SELF extends ValidationAssert<SELF, VALIDATION, T>
      * Asserts that the validation is invalid.
      *
      * @param actual the actual value.
-     
+
      * @return an {@link InvalidValidationAssert} for further assertions.
      */
     public static <T> InvalidValidationAssert<?, Validation.Invalid<T>, T> assertInvalid(Validation<T> actual) {
@@ -77,6 +78,18 @@ public class ValidationAssert<SELF extends ValidationAssert<SELF, VALIDATION, T>
         Validation<T> validation = Validation.from().catching(codeThrowingValidationException);
         assertThat(validation.isInvalid()).as("Expected validation to be invalid but was valid").isTrue();
         return new InvalidValidationAssert<>((Validation.Invalid<T>) validation);
+    }
+
+    /**
+     * Asserts that code passed will throw a ValidationException.
+     */
+    public static InvalidValidationAssert<?, Validation.Invalid<Object>, Object> assertInvalid(Runnable codeThrowingValidationException) {
+        Validation<?> validation = Validation.from().catching(() -> {
+            codeThrowingValidationException.run();
+            Assertions.fail("Expected codeThrowingValidationException to throw ValidationException");
+            return null;
+        });
+        return new InvalidValidationAssert<>((Validation.Invalid<Object>) validation);
     }
 
 }